renamed more files to use correct extensions
