+++ pages/tools.tsx
